Add files via upload
Usually Gini score based feature selection is common for random forest models but such model will take too much run time and is not effective when the data is huge. In such a case, ranger function is very effective which is just a faster way to create random forest models.

Please note, as a prerequisite, there shouldn't be any missing values in the data so we would have to remove it before this method.