Add files via upload
Problem - Common problem that we see after creating models is their propensity to favor a particular result let's say class A compared to class B. This is a very common problem which occurs due to high number of rows of class A compared to class B in the training data. Therefore, the model created will have higher error rate for class B compared to class A.

Solution - To avoid such problem, we scale the training data in such a manner that class A and class B have equal rows. In this code, we demonstrate two types of functions:
1) downscale - Decreases the instances of the majority class to equal the minority class.
2) upscale - Increases the instances of the minority class by repeating instances to equal the majority class.