Add files via upload
I recently worked on a basic deep learning project which involved recognizing a digit [0-9] from a handwritten image through neural networks. This has potential applications in image to text conversion specially while automating historical forms/records. Input image consisted of 784 pixels of a handwritten number. To start with the problem, we sampled the data to have equal training datasets for each digit based on varying handwriting styles. We then normalized 784 pixels of each input image from [0,255] to [0,1] for better training of the model. Upon varying the hidden layers and neurons, we were able to achieve 97.5% overall accuracy with output range - 96% for digit 3 and 99.38% for digit 0. This was somewhat expected since digit 3 is similar to 8 while 0 is a unique digit.